去除一些编译警告

+++ test/util/txtcfg/xml/test_xml_parser.cpp
@@ -15,7 +15,7 @@ string tag;
 
 class Handle : public XmlElementHandler
 {
-    
+
 public:
     virtual void handle_attribute(const std::string &name, const std::string &value)
     {
@@ -23,21 +23,21 @@ public:
         if (name == "a" && value == "b&\"<>")
             tag += "a";
     }
-    
+
     virtual void handle_text(const std::string& text)
     {
         //printf("text %s\n", text.data());
         if (text == "c")
             tag += "t";
     }
-    
+
     virtual void handle_comment(const std::string& comment)
     {
         //printf("comment %s\n", comment.data());
         if (comment == "d")
             tag += "c";
     }
-    
+
     virtual XmlElementHandler* handle_child(const std::string &name)
     {
         //printf("child %s\n", name.data());
@@ -45,21 +45,21 @@ public:
             tag += "h";
         return this;
     }
-    
+
     virtual void handle_child_finish(XmlElementHandler *child)
     {
         //printf("fin child\n");
         if (child == this)
             tag += "x";
     }
-    
+
     virtual void handle_finish()
     {
         //printf("fin\n");
         tag += "y";
     }
 };
-    
+
 }
 
 NUT_FIXTURE(TestXmlParser)
@@ -74,7 +74,7 @@ NUT_FIXTURE(TestXmlParser)
     }
 
     void tearDown() {}
-    
+
     void testSmoke()
     {
         Handle h;
@@ -84,7 +84,7 @@ NUT_FIXTURE(TestXmlParser)
         p.input(s);
         p.finish();
         if (p.has_error())
-            printf("xml error %d:%d %s\n", p.line(), p.column(), p.error_message().data());
+            printf("xml error %d:%d %s\n", (int)p.line(), (int)p.column(), p.error_message().data());
         NUT_TA(!p.has_error());
         //printf("%s\n", tag.data());
         NUT_TA(tag == "hayxtcyx");
